add ProceduresAgent class
